refactor: centralize file utilities
