+++ indexManager.js
@@ -3,48 +3,16 @@ const path = require('path');
 const github = require('./githubClient');
 const tokenStore = require('./tokenStore');
 const memoryConfig = require('./memoryConfig');
+const {
+  ensureDir,
+  deepMerge,
+  normalizeMemoryPath,
+  generateTitleFromPath,
+  inferTypeFromPath
+} = require('./utils/fileUtils');
 
 const indexPath = path.join(__dirname, 'memory', 'index.json');
 let indexData = null;
-
-function isObject(val) {
-  return val && typeof val === 'object' && !Array.isArray(val);
-}
-
-function deepMerge(target, source, matchKey) {
-  if (Array.isArray(target) && Array.isArray(source)) {
-    const result = [...target];
-    source.forEach(item => {
-      if (matchKey && isObject(item)) {
-        const idx = result.findIndex(e => isObject(e) && e[matchKey] === item[matchKey]);
-        if (idx >= 0) {
-          result[idx] = deepMerge(result[idx], item, matchKey);
-        } else {
-          result.push(item);
-        }
-      } else if (!result.includes(item)) {
-        result.push(item);
-      }
-    });
-    return result;
-  } else if (isObject(target) && isObject(source)) {
-    const out = { ...target };
-    Object.keys(source).forEach(k => {
-      if (k in target) {
-        out[k] = deepMerge(target[k], source[k], matchKey);
-      } else {
-        out[k] = source[k];
-      }
-    });
-    return out;
-  }
-  return source;
-}
-
-function ensureDir(filePath) {
-  const dir = path.dirname(filePath);
-  if (!fs.existsSync(dir)) fs.mkdirSync(dir, { recursive: true });
-}
 
 function readLocalIndex() {
   if (!fs.existsSync(indexPath)) return [];
@@ -58,35 +26,7 @@ function readLocalIndex() {
   }
 }
 
-function normalizeMemoryPath(p) {
-  if (!p) return 'memory/';
-  let rel = p.replace(/\\+/g, '/');
-  rel = path.posix.normalize(rel).replace(/^(\.\/)+/, '').replace(/^\/+/, '');
-  while (rel.startsWith('memory/')) {
-    rel = rel.slice('memory/'.length);
-  }
-  return path.posix.join('memory', rel);
-}
 
-function generateTitleFromPath(p) {
-  return p
-    .split('/')
-    .pop()
-    .replace(/\..+$/, '')
-    .replace(/[-_]/g, ' ')
-    .replace(/\b\w/g, l => l.toUpperCase());
-}
-
-const { detectMarkdownCategory } = require('./markdownCategory');
-
-function inferTypeFromPath(p) {
-  if (p.endsWith('.md')) return detectMarkdownCategory(p);
-  if (p.includes('plan')) return 'plan';
-  if (p.includes('profile')) return 'profile';
-  if (p.includes('lesson')) return 'lesson';
-  if (p.includes('note')) return 'note';
-  return 'file';
-}
 
 async function githubWriteFileSafe(token, repo, relPath, data, message, attempts = 2) {
   for (let i = 1; i <= attempts; i++) {
